+++ api/src/modules/document/resolvers.ts
@@ -1,6 +1,7 @@
 import { interfaceType, list, nonNull, objectType, queryField } from 'nexus';
 import { Context } from '../../graphql/server/drive/context';
 import { getChildLogger } from '../../logger';
+import { OperationContext } from '../document-drive/drive-resolver';
 import { GQLDateBase } from '../system';
 
 const logger = getChildLogger({ msgPrefix: 'DOCUMENT RESOLVER' });
@@ -17,10 +18,9 @@ export const operationModelInterface = interfaceType({
     t.string('id');
     t.nonNull.string('inputText');
     t.string('error');
+    t.field('context', { type: OperationContext });
   },
-  resolveType: e => {
-    return 'DefaultOperation';
-  }
+  resolveType: () => null
 });
 
 export const operationModel = objectType({
@@ -40,7 +40,28 @@ export const documentModelInterface = interfaceType({
     t.nonNull.int('revision');
     t.nonNull.field('created', { type: GQLDateBase });
     t.nonNull.field('lastModified', { type: GQLDateBase });
-    t.nonNull.list.nonNull.field('operations', { type: operationModel });
+    t.nonNull.list.nonNull.field('operations', {
+      args: { first: 'Int', skip: 'Int' },
+      type: operationModel,
+      resolve: async (parent, { first, skip }, ctx: Context) => {
+        if (!ctx.driveId) {
+          throw new Error('DriveId is not defined');
+        }
+        const document =
+          ctx.documents.get(parent.id) ??
+          (await ctx.prisma.document.getDocument(ctx.driveId, parent.id));
+        const { operations } = document;
+        if (first && skip) {
+          return operations.slice(skip, skip + first);
+        }
+
+        if (first) {
+          return operations.slice(0, first);
+        }
+
+        return operations;
+      }
+    });
   },
   resolveType: e => {
     switch (e.documentType) {
@@ -75,20 +96,25 @@ export const documentQuery = queryField('document', {
       throw new Error('DriveId is not defined');
     }
     const doc = await ctx.prisma.document.getDocument(ctx.driveId, id);
+    ctx.documents.set(doc.id, doc);
     return doc;
   }
 });
 
 export const documentsQuery = queryField('documents', {
   type: list(documentModelInterface),
-  resolve: async (_root, { id }, ctx: Context) => {
+  resolve: async (_root, args, ctx: Context) => {
     if (!ctx.driveId) {
       throw new Error('DriveId is not defined');
     }
     try {
       const docIds = await ctx.prisma.document.getDocuments(ctx.driveId);
       const docs = await Promise.all(
-        docIds.map(doc => ctx.prisma.document.getDocument(ctx.driveId!, doc))
+        docIds.map(async id => {
+          const doc = await ctx.prisma.document.getDocument(ctx.driveId!, id);
+          ctx.documents.set(doc.id, doc);
+          return doc;
+        })
       );
       return docs;
     } catch (e: any) {